requirements.txt
glfw==1.11.0
PyGLM==1.1.8
PyOpenGL==3.1.5
